[Unphased] Moving app to Examples folder
