--- DropBearApp/AppDelegate.swift
@@ -1,14 +1,0 @@
-import UIKit
-
-class AppDelegate: UIResponder, UIApplicationDelegate {
-    var window: UIWindow?
-
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        let window = UIWindow(frame: UIScreen.main.bounds)
-        let nav = UINavigationController(rootViewController: ScrollableStackViewController())
-        window.rootViewController = nav
-        window.makeKeyAndVisible()
-        self.window = window
-        return true
-    }
-}
