+++ Examples/DropBearApp/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
