+++ Examples/DropBearApp/Resources/Assets.xcassets/Contents.json
