+++ Examples/DropBearApp/Screens/TableView/TableViewCell.xib
