+++ Examples/DropBearApp/UITestSceneDelegate.swift
@@ -1,0 +1,7 @@
+import UIKit
+
+class UITestSceneDelegate: SceneDelegate {
+    func sceneDidBecomeActive(_ scene: UIScene) {
+        window?.layer.speed = 100
+    }
+}
