+++ Examples/DropBearAppUITests/Robots/ScrollableStackRobot.swift
